Clean up by moving check_operation into own file
